move enterprise repo only if exists

+++ scripts/apply.sh
@@ -7,7 +7,7 @@ FREE_REPO_LINE="deb http://download.proxmox.com/debian/pve `lsb_release -cs` pve
 function pve_patch() {
   echo "apply patch..."
   echo $FREE_REPO_LINE > $FREE_REPO_LIST
-  mv $ENTERPRISE_REPO_LIST $ENTERPRISE_REPO_LIST~
+  [ -f $ENTERPRISE_REPO_LIST ] && mv $ENTERPRISE_REPO_LIST $ENTERPRISE_REPO_LIST~
   sed -i.bak "s/data.status !== 'Active'/false/g" /usr/share/javascript/proxmox-widget-toolkit/proxmoxlib.js
   cp --backup /usr/share/pve-patch/images/* /usr/share/pve-manager/images/
 }
